+++ lib/core/io/domain/interfaces/i_manage_files_service.dart
