+++ lib/core/io/domain/interfaces/i_permissions_service.dart
